requirements.dev.txt
argon2-cffi==18.1.0
cffi==1.11.5
Django==2.0.2
django-debug-toolbar==1.9.1
django-template-debug==0.3.5
pkg-resources==0.0.0
psycopg2-binary==2.7.4
pycparser==2.18
python-decouple==3.1
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
